+++ src/components/OutsideClickHandler.jsx
@@ -36,7 +36,7 @@ export default class OutsideClickHandler extends React.Component {
   }
 
   onOutsideClick(e) {
-    const isDescendantOfRoot = this.childNode.contains(e.target);
+    const isDescendantOfRoot = this.childNode != null && this.childNode.contains(e.target);
     if (!isDescendantOfRoot) {
       this.props.onOutsideClick(e);
     }